Removes extra requirements.
