+++ setup.py
@@ -10,7 +10,6 @@ install_requirements = [
     'imageio', 'torch', 'imageio', 'matplotlib', 'progressbar2', 'scipy',
     'sklearn', 'torchvision', 'visdom', 'pyyaml', 'pathlib'
 ]
-extras_requirements = {'docs':['sphinx >= 1.4','sphinx_rtd_theme', 'numpydoc']}
 
 setup(
     name='cortex',
@@ -20,7 +19,5 @@ setup(
     author_email='erroneus@gmail.com',
     packages=packages,
     install_requires=install_requirements,
-    extras_require=extras_requirements,
     entry_points={'console_scripts': ['cortex=cortex.main:main']},
     zip_safe=False)
-
